nuvens, sol, e lua

+++ gdx-jogo/src/com/minimine/cenas/Bloco.java
@@ -2,25 +2,25 @@ package com.minimine.cenas;
 
 public class Bloco {
 	public CharSequence nome;
-	public byte tipo;
+	public int tipo;
 	public int topo, lados, baixo;
 	public boolean transparente = false;
 	public boolean solido = true;
 	
-	public Bloco(CharSequence nome, byte tipo, int topo) {
+	public Bloco(CharSequence nome, int tipo, int topo) {
 		this.nome = nome;
 		this.tipo = tipo;
 		this.topo = topo; this.lados = topo; this.baixo = topo;
 	}
 	
-	public Bloco(CharSequence nome, byte tipo, int topo, boolean transparente) {
+	public Bloco(CharSequence nome, int tipo, int topo, boolean transparente) {
 		this.nome = nome;
 		this.tipo = tipo;
 		this.topo = topo; this.lados = topo; this.baixo = topo;
 		this.transparente = transparente;
 	}
 	
-	public Bloco(CharSequence nome, byte tipo, int topo, boolean transparente, boolean solido) {
+	public Bloco(CharSequence nome, int tipo, int topo, boolean transparente, boolean solido) {
 		this.nome = nome;
 		this.tipo = tipo;
 		this.topo = topo; this.lados = topo; this.baixo = topo;
@@ -28,14 +28,14 @@ public class Bloco {
 		this.solido = solido;
 	}
 	
-	public Bloco(CharSequence nome, byte tipo, int topo, int lados) {
+	public Bloco(CharSequence nome, int tipo, int topo, int lados) {
 		this.nome = nome;
 		this.tipo = tipo;
 		this.topo = topo;
 		this.lados = lados; this.baixo = topo;
 	}
 	
-	public Bloco(CharSequence nome, byte tipo, int topo, int lados, int baixo) {
+	public Bloco(CharSequence nome, int tipo, int topo, int lados, int baixo) {
 		this.nome = nome;
 		this.tipo = tipo;
 		this.topo = topo;
